+++ src/main/java/com/github/daniel_sc/rocketchat/modern_client/response/DateWrapper.java
@@ -7,4 +7,11 @@ import java.util.Date;
 public class DateWrapper {
     @SerializedName("$date")
     public Date date;
+
+    @Override
+    public String toString() {
+        return "DateWrapper{" +
+                "date=" + date +
+                '}';
+    }
 }