+++ src/main/java/com/github/daniel_sc/rocketchat/modern_client/response/livechat/Condition.java
@@ -1,0 +1,14 @@
+package com.github.daniel_sc.rocketchat.modern_client.response.livechat;
+
+public class Condition {
+    public String name;
+    public String value;
+
+    @Override
+    public String toString() {
+        return "Condition{" +
+                "name='" + name + '\'' +
+                ", value='" + value + '\'' +
+                '}';
+    }
+}