+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Expose all raw messages - see #23, #24.
+
 ## [1.2.0] - 2021-10-04
 
 ### Added